requirements.txt
certifi
pyreadline3 ~= 3.4.0; platform_system == 'Windows'
httpx[socks]~=0.23.3
requests[socks]~=2.28.2
rich~=13.3.2
appdirs~=1.4.4
werkzeug~=2.2.3
flask~=2.2.3
flask-cors~=3.0.10
loguru~=0.6.0
sentry-sdk~=1.17.0
pyjwt[crypto]~=2.6.0
pyperclip~=1.8.2
pandora-chatgpt~=1.1.5
